reqif: reqif-to-sdoc: remove the factory class to simplify further work
Also: tests/integration: ReqIF example produced by Polarion (disabled for now)
