+++ pyproject.toml
@@ -59,7 +59,7 @@ dependencies = [
     "XlsxWriter >= 1.3.7, == 1.*",
     "xlrd >= 2.0.1, == 2.*",
 
-    "reqif >= 0.0.23, == 0.*",
+    "reqif >= 0.0.25, == 0.*",
 
     # Bibtex
     "pybtex >= 0.23.0, == 0.*",
