+++ tests/integration/backend/reqif/native/end_to_end/40_specification_type/test.itest
@@ -3,8 +3,7 @@ RUN: %cat %S/output/reqif/output.reqif | filecheck %s
 RUN: %strictdoc import reqif sdoc %S/output/reqif/output.reqif %S/output/sample.sdoc
 RUN: %diff %S/sample.sdoc %S/output/sample.sdoc
 
-CHECK:        <SPECIFICATION-TYPE IDENTIFIER="SDOC_SPECIFICATION_TYPE_SINGLETON" LAST-CHANGE="{{.*}}" LONG-NAME="SDOC_SPECIFICATION_TYPE_SINGLETON">
-CHECK:        </SPECIFICATION-TYPE>
+CHECK:        <SPECIFICATION-TYPE IDENTIFIER="SDOC_SPECIFICATION_TYPE_SINGLETON" LAST-CHANGE="{{.*}}" LONG-NAME="SDOC_SPECIFICATION_TYPE_SINGLETON"/>
 CHECK:        <SPECIFICATION IDENTIFIER="SPECIFICATION-{{.*}}" LONG-NAME="Document one requirement">
 CHECK:          <TYPE>
 CHECK:            <SPECIFICATION-TYPE-REF>SDOC_SPECIFICATION_TYPE_SINGLETON</SPECIFICATION-TYPE-REF>
