+++ tests/integration/backend/reqif/native/examples/02_Polarion/test.itest
@@ -1,0 +1,18 @@
+UNSUPPORTED: true
+
+RUN: %mkdir %S/output
+RUN: %strictdoc import reqif sdoc %S/sample.reqif %S/output/output.sdoc
+UN: %cat %S/output/output.sdoc | filecheck %s
+
+CHECK: [REQUIREMENT]
+CHECK: UID: Anonymized-194ac1eb-d718-b59b-5d2a-c2b0ac0381ef
+CHECK: STATEMENT: >>>
+CHECK: <xhtml:div>...Anonymized...</xhtml:div>
+CHECK: <<<
+CHECK: COMMENT: >>>
+CHECK: Anonymized-6e95e816-6fc9-0fe1-8b1f-66340311da94
+CHECK: <<<
+
+RUN: %strictdoc export --formats=reqif-sdoc %S/output/output.sdoc
+UN: %strictdoc import reqif sdoc %S/output/reqif/output.reqif %S/output/output_step2.sdoc
+UN: %diff %S/output/output.sdoc %S/output/output_step2.sdoc
